Moved the notes file in /pages/Notes
